refactor(routines): split run-listing out of service_trigger.rs
Auto-pull's addition pushed service_trigger.rs past the repo's
500-line-per-file cap. Move svc_list_runs/svc_list_all_runs and their
shared run_summary helper into a new sibling service_run_list.rs —
pure extraction, no behavior change, both files stay under the cap.
